fix(front): refactor
